refactored a lot as per feedback receieved
convert-hex-vector/convert-vector-int are now more elegant. they use a
lot less converting to/from strings. similarly, oid-str now populates a
full-length string beforehand and sets arefs into in instead of (setf
str (concatenate 'string str ...)) which is a but more sensible.

get-current-pid now takes :if-not-exists-return keyword arg, which
returns a value when we're unable to pull PID from system.